Use conditional shallow/deep wrapping.
